backend/requirements.txt
# Backend dependencies
fastapi
uvicorn
langgraph
pydantic
typing_extensions
# langchain_openai  # Replaced with Ollama
langchain_core
langchain_community
langchain-ollama
python-dotenv
sse-starlette
pyppeteer
langchain-mcp-adapters
pandas
numpy
matplotlib
seaborn
python-multipart